--- star_rail/core/__init__.py
@@ -1,2 +1,0 @@
-from .async_request import *
-from .paginator import *
