+++ star_rail/module/account/__init__.py
@@ -1,1 +1,2 @@
-from .account import *
+from .account import Account
+from .client import *
