+++ star_rail/module/helper/__init__.py
@@ -1,0 +1,3 @@
+from .header import *
+from .paginator import *
+from .request import *
