+++ star_rail/module/helper/paginator/__init__.py
