+++ star_rail/module/helper/paginator/api.py
