+++ star_rail/module/helper/paginator/base.py
