+++ star_rail/module/month/client.py
@@ -1,11 +1,11 @@
 import typing
 
 from star_rail import exceptions as error
-from star_rail.core import request
 from star_rail.module import routes
 from star_rail.module.base import BaseClient
 from star_rail.utils.logger import logger
 
+from ..helper import request
 from .mapper import MonthInfoItemMapper
 from .model import MonthInfoData, MonthInfoItem
