+++ star_rail/module/updater.py
@@ -1,7 +1,7 @@
+import aiohttp
 import yarl
 
 from star_rail import __version__ as app_version
-from star_rail.core import _request
 from star_rail.utils.logger import logger
 from star_rail.utils.version import compare_versions
 
@@ -17,7 +17,10 @@ class Updater:
         """
         logger.debug("Check update.")
         url = yarl.URL("https://api.github.com/repos/cntvc/star-rail-tools/releases/latest")
-        data = await _request("GET", url)
+
+        async with aiohttp.ClientSession() as session:
+            async with session.request(method="GET", url=url) as response:
+                data = await response.json()
 
         latest_version = data["tag_name"]
         return compare_versions(app_version, latest_version) == -1
